[Runner] Small code improvements, fixed some warnings, enabled -Wextra

+++ runner/src/FileZip.cpp
@@ -12,7 +12,7 @@ void FileZip::decodeEntry(size_t entry_number, zip_stat_t *entry_stats) {
   std::string entry_name(entry_stats->name);
   std::vector<char> data_buffer;
 
-  size_t read_bytes = 0;
+  int64_t read_bytes = 0;
 
   zip_file_t *entry = NULL;
